requirements.txt
fastapi
uvicorn
psycopg[binary]
chromadb
pypdf2
python-docx
python-jose[cryptography]
passlib[bcrypt]
openpyxl
python-multipart
requests
aiofiles
httpx
sentence-transformers
langchain
transformers
torch
nltk